feat: :recycle: Split explore page into components
Removed nested components inside explore page to designated files.
