+++ src/routes/explore.tsx
@@ -1,5 +1,5 @@
 import { createFileRoute } from "@tanstack/react-router";
-import Explore from "../components/Explore";
+import Explore from "../components/explore/Explore";
 
 export const Route = createFileRoute("/explore")({
   component: Explore,
